Avoid MissingH, use 'split' instead
The MissingH package by John Goerzen has been on life support for years.
It was maintained by hvr, but he is no longer active. The package relies
on the Hackage trustees to get its bounds bumped.

It is also a potpourri, and we really only to replace within a string.

In contrast, the split package is maintained by Brent Yorgey, which is
still active in the Haskell community. And it is only focused on
splitting lists. It also has no dependencies.

All of these advantages must be weighed against the fact that we now
need to define 'replace' ourselves using the building blocks from
'split'. I think that the advantages are overwhelming.